phase4_projects/03_code_review_agent/requirements.txt
# ==================== Web Framework ====================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
requests>=2.31.0
python-multipart>=0.0.6
sse-starlette>=1.6.0

# ==================== LangChain & LangGraph ====================
langchain>=0.1.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langgraph>=0.0.20

# ==================== MCP Integration ====================
langchain-mcp-adapters>=0.1.0
mcp>=1.0.0

# ==================== Vector Database ====================
chromadb>=0.5.0

# ==================== Git Operations ====================
gitpython>=3.1.0

# ==================== Document Processing ====================
pypdf>=5.0.0
python-docx>=1.1.0
PyYAML>=6.0.0

# ==================== Embeddings ====================
langchain-huggingface>=0.1.0
sentence-transformers>=2.2.0

# ==================== Data & Validation ====================
pydantic>=2.5.0
pydantic-settings>=2.1.0
numpy>=1.24.0

# ==================== Environment ====================
python-dotenv>=1.0.0
aiofiles>=23.2.0
httpx>=0.25.0
